feat(go_router): add refreshListenable for dynamic routing
Implement `refreshListenable` in the GoRouter configuration using a
`RouterRefreshNotifier`. This allows the router to react to stream
events, enabling dynamic changes to the navigation state.

A new `RouterRefreshNotifier` class, which extends `ChangeNotifier`, is
introduced to listen to a stream and notify listeners upon receiving
new data. This is demonstrated by triggering route refreshes from a
stream in `MaterialGoShellRouteTest1`.